internal/helm: add repository cache helpers
This commits adds simple caching capabilities to the
`ChartRepository`, which makes it possible to load the `Index` from a
defined `CachePath` using `LoadFromCache()`, and to download the index
to a new `CachePath` using `CacheIndex()`.

In addition, the repository have been updated to make use of Gomega
and missing ones have been added.

Signed-off-by: Hidde Beydals <hello@hidde.co>