feat: reconcile HelmCharts sourced from OCIRepositories
The HelmChartReconciler is now able to reconcile HelmCharts that have
an OCIRepository source. For now all it does is pull the chart and
store it as an artifact. It does not support all the fields of
HelmCharts such as ValuesFile/ValuesFiles.

I changed the OCIRepositoryReconciler so that it pulls a reference's
manifest and stores the revision as well as the URL in the
OCIRepository's status.

Signed-off-by: Max Jonas Werner <mail@makk.es>